Exceptions in service layer

+++ formula1-service/src/main/java/cz/muni/pa165/teamwhite/formula1/service/ComponentServiceImpl.java
@@ -2,9 +2,16 @@ package cz.muni.pa165.teamwhite.formula1.service;
 
 import cz.muni.pa165.teamwhite.formula1.persistence.dao.ComponentDao;
 import cz.muni.pa165.teamwhite.formula1.persistence.entity.Component;
+import cz.muni.pa165.teamwhite.formula1.service.exception.Formula1ServiceException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
 
 import java.util.List;
+
+/**
+ * @author Tomas Sedlacek
+ */
+
 
 public class ComponentServiceImpl implements ComponentService{
     @Autowired
@@ -12,31 +19,57 @@ public class ComponentServiceImpl implements ComponentService{
 
     @Override
     public Long createComponent(Component component) {
-        componentDao.create(component);
+        try {
+            componentDao.create(component);
+        } catch (DataAccessException e) {
+            throw new Formula1ServiceException("Could not create component: " + component, e);
+        }
 
         return component.getId();
     }
 
     @Override
     public void update(Component component) {
-        componentDao.update(component);
+        try {
+            componentDao.update(component);
+        } catch (DataAccessException e) {
+            throw new Formula1ServiceException("Could not update component: " + component, e);
+        }
     }
 
     @Override
     public List<Component> findAll() {
-        return componentDao.findAll();
+        List<Component> result;
+
+        try {
+            result = componentDao.findAll();
+        } catch (DataAccessException e) {
+            throw new Formula1ServiceException("Could not return all components", e);
+        }
+        return result;
     }
 
     @Override
     public Component findById(Long id) {
-        return componentDao.findById(id);
+        Component found;
+
+        try {
+            found = componentDao.findById(id);
+        } catch (DataAccessException e) {
+            throw new Formula1ServiceException("Could not find component with ID: " + id, e);
+        }
+
+        return found;
     }
 
     @Override
     public void remove(Long id) {
-        Component toDelete = componentDao.findById(id);
-        if (toDelete != null){
-            componentDao.remove(toDelete);
+        Component toDelete = findById(id);
+
+        try {
+            if (toDelete != null) componentDao.remove(toDelete);
+        } catch (DataAccessException e) {
+            throw new Formula1ServiceException("Could not remove component with ID: " + id, e);
         }
     }
 }
